requirements.txt
google-generativeai
python-dotenv
langchain
faiss-cpu
langchain_google_genai
flask